Test #4 <Card /> component
enabling testing on the card component while deleting the onclick event on the <Item />
component as it was making the test to throw an error.

Note: The <Card /> component isn't fully tested yet, there still are fixes that need to
be done on the <Card /> component and the test itself.