Fix Spinner over user's profile image
This fix keeps spinning the spinner until firebase returns some data to show on the
cards, before the spinner had a static time to spin over the page but now the page is
locked for as long as we come up with data to show on the screen. Here data is the
content (series, films).